refactor: constants 폴더 이동 및 index에서 export
