+++ src/constants/index.js
@@ -1,0 +1,4 @@
+const MESSAGE = require('./message');
+const RULE = require('./rule');
+
+module.exports = { ...MESSAGE, ...RULE };
